Update combat scripts to be generalized
