+++ HackUSU22/Assets/Scripts/Combat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d445eb08c9758a44846a28662a20551
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
