+++ HackUSU22/Assets/Scripts/EnemyCombat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 163577759703e7e4d9528287a4ba52d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
